feat(tech-tree): Integrate International Trade upgrade
This commit integrates the `InternationalTrade` upgrade into the research tree.

The `InternationalTrade` upgrade was already defined as an `UpgradeType` and had a cost associated with it in the configuration, but it was not available to be researched in the tech tree.

This commit makes the following changes:

- In `src/core/tech/TechEffects.ts`:
  - Adds `INTERNATIONAL_TRADE: Economy-2` to the `RESEARCH_TECH_IDS` constant.
  - Adds a new `TechDefinition` for `INTERNATIONAL_TRADE` to the `TECHS` constant, which unlocks the `UpgradeType.InternationalTrade` upgrade upon completion.

This change makes the `InternationalTrade` upgrade available to be researched at the Economy-2 position in the tech tree.